+++ src/core/UpdateEmployee.java
@@ -243,8 +243,8 @@ public class UpdateEmployee extends JFrame implements ActionListener {
                         employee.setDOB(calendar);
                         employee.setContactNum(txtContactNum.getText());
                         employee.setEmail(txtEmail.getText());
-                        employee.setNumHolidays(Double.valueOf(txtNumHoldiays.getText()));
-                        employee.setContractHours(Double.valueOf(txtContractHours.getText()));
+                        employee.setNumHolidays(Integer.valueOf(txtNumHoldiays.getText()));
+                        employee.setContractHours(Integer.valueOf(txtContractHours.getText()));
                         employee.setSalary(Double.valueOf(txtSalary.getText()));
                         employee.setOnHoliday(Double.valueOf(txtOnHoliday.getText()));
                         employee.setOffSick(Double.valueOf(txtOffSick.getText()));
